ticket cannot be deleted less than 4 hours before screening

+++ Jegymester.ServicesA/TicketService.cs
@@ -41,11 +41,33 @@ namespace Jegymester.Services
                 throw new KeyNotFoundException("Ticket not found.");
             }
 
+            //specifikáció: 4 órával a vetítés előtt már nem törölhető a jegy, de addig igen
+            /*
+             * ötlet:
+             * 0) új változó: 4 órát levonunk a screeningstarttimeból
+             * 1) megnézzük hogy a datetime.now nagyobb e mint az új változó
+             * 2) ha igen, akkor exceptiont dobunk
+             */
+
+            // 0)
+            TimeSpan converted_time = TimeSpan.FromHours(4);
+            DateTime deadline = ticket.Screening.ScreeningStartTime - converted_time;
+            // 1)
+            if (DateTime.Now > deadline) 
+            {
+                //2)
+                throw DeadlineException("Ticket cannot be deleted because there is less than 4 hours until the screening");
+            }
+
             _context.Tickets.Remove(ticket);
             await _context.SaveChangesAsync();
             return true;
         }
 
+        private Exception DeadlineException(string v)
+        {
+            throw new NotImplementedException();
+        }
     }
 }
 /* ============================================= UNDER DEV ========================================= */
